uses eff. ion parameter for ion grad model
From now on we use the effective ionization parameter for
the defintion of the ion grad alpha model. It is defined
as xi_eff = xi * cos(PI/4) / cos(del_inc).

This change required the interpolation routine of the table to only
accept positive numbers for the incident angle (without loss of
information relevant for the model).