Merge branch 'rrad-remove-return' into master
This merge commit was a substantial effort to combine all changes from
the return radiation model branch into the main master branch (without
the model itself)

Merging the "Mega-Merge" of getting the changes from the return
radiation commit into the master. Fixes are applied and all test
scripts working correctly.

Caveat: if something used to work before this commit, but does not any
more now, this commit is likely to be the problem

 Conflicts:
        CMakeLists.txt
        Makefile
        common.h
        relbase.c
        relbase.h
        relcache.c
        rellp.c
        rellp.h
        relmodels.c
        relmodels.h
        reltable.c
        relutility.c
        relutility.h
        test/e2e/check_model_functions.sl
        test_relxill.c
        test_relxill.h
        test_sta.c
        test_std_functions.c
        test_xilltab.c
        xilltable.c