fixes treatment of gshift-fluxcorr
Now for g<1 no values of the fluxcorrection >1 are used. Before that,
the code tried to enforce this by an assert, which sometimes failed
and crashed the code.

Detailed investigation showed, that in those cases the flux correction
was less than 1% above 1, so negligible. If it is >1%, a warning is issed.